Changes are:
1)Rashid's fix in gfs dynamics gfidi_hyb_gc_h.f, it adds p0 as an correction to compute the terms in vertical advection in thermodynamics equation.

2)FOR WAM, tmp is updated after radiation and idea physics, before gbphys is called.

3)A scale parameter cleff in gravity wave drag (gwdps.f) is set to be same as operational setting, this will change gfs results and NMMB with gfs physics results.

4)Moorthi's bug fixes in writing out gfluxi,ice thickness, and snow depth in flx file are merged to trunk.

5)path for stmp when generation new baseline is fixed in rt_gfs.sh.

In those changes, 1) and 4) are done in gsm branch,see gsm ticket:
 https://svnemc.ncep.noaa.gov/trac/gsm/ticket/19

Due to change 3) and 4), gfs regression test results are changed. And due to change 3), NMMB run with gfs physics regression test results are changed. New regression test results are generated on ccs, zeus and wcoss. Full regression test passed on all three machines.

To view the code:

on ccs:
/climate/save/wx20wa/wam/nems/wamfix_gfidigwd
on tide:
/nems/save/Jun.Wang/nems/wam/wamfix_gfidigwd
on zeus:
/scratch1/portfolios/NCEPDEV/nems/save/Jun.Wang/wam/nems/wamfix_gfidigwd


git-svn-id: https://svnemc.ncep.noaa.gov/projects/nems/trunk@28991 281fb692-3a18-0410-826d-9e847983f365