Changes are:
Current NEMS code can restart from history files, but digital filter does not work for this case. The time in GFS clock/alarm is updated to allow digital filter work for restart
2. Ruiyu found that the sigma level used for cloud initialization needs vertical coordicate information ak5 and bk5, but current gsm phys does not have ak5,bk5, instead, it uses the ratio of vertical level to compute sigma level, which causes reduce PBL clouds. A subroutine is added to read the vertical coordinate from grid point model level file, and used to compute the sigma level for cloud initialization.

3. high frequency output is limited only for cold start

NEMS ticket is:

https://svnemc.ncep.noaa.gov/trac/nems/ticket/152

corresponding gsm ticket is:

https://svnemc.ncep.noaa.gov/trac/gsm/ticket/82

The nems branch is at:

https://svnemc.ncep.noaa.gov/projects/nems/branches/jun/dfi_rsthst

The gsm regression test results are generated, the regression test passed on both theia and wcoss. There is no impact on nmm.



git-svn-id: https://svnemc.ncep.noaa.gov/projects/nems/trunk@71595 281fb692-3a18-0410-826d-9e847983f365